images/ci-prune-buildcache/requirements.txt
awscli
boto3
python-gitlab
sentry-sdk
